Reference

Abstract types

SimpleDirectedHypergraphs.AbstractDirectedHypergraph — Type

AbstractDirectedHypergraph{T} <: AbstractHypergraph{T}

An abstract directed hypergraph type storing information about vertices and dihyperedges.

Creating a directed hypergraph

SimpleDirectedHypergraphs.DirectedHypergraph — Type

DirectedHypergraph{T} <: AbstractDirectedHypergraph{Tuple{Union{T, Nothing}, Union{T, Nothing}}}

A directed hypergraph (dihypergraph) storing information about vertices and directed hyperedges (dihyperedges).

This implementation is based on guidance from Przemysław Szufel; see https://github.com/pszufe/SimpleHypergraphs.jl/issues/45 This allows us to manipulate DirectedHypergraphs using Hypergraph functionality There is danger of a user manipulating individual hg_tail and hg_head (undirected) hypergraphs

Is there a smart way to prevent this? TODO: reconsider this design choice

Constructors

DirectedHypergraph{T,V,E,D}(
    n::Integer, k::Integer,
    v_meta=Vector{Union{V, Nothing}}(nothing, n),
    he_meta_tail=Vector{Union{E, Nothing}}(nothing, k),
    he_meta_head=Vector{Union{E, Nothing}}(nothing, k)
) where {T<:Real,V,E,D<:AbstractDict{Int, T}}
DirectedHypergraph{T,V,E}(n::Integer, k::Integer) where {T<:Real, V, E}
DirectedHypergraph{T,V}(n::Integer, k::Integer) where {T<:Real, V}
DirectedHypergraph{T}(n::Integer, k::Integer) where {T<:Real}
DirectedHypergraph(n::Integer, k::Integer)

Construct a dihypergraph with a given number of vertices and hyperedges. Optionally, values of type V can be stored at vertices and values of type E can be stored at hyperedges. By default the dihypergraph uses a Dict{Int,T} for the internal data storage; however, a different dictionary such as SortedDict to ensure result replicability can be used (e.g., when doing stochastic simulations on dihypergraphs).

DirectedHypergraph(
    m_tail::AbstractMatrix{Union{T, Nothing}},
    m_head::AbstractMatrix{Union{T, Nothing}}
) where {T<:Real}    
DirectedHypergraph{T}(
    m_tail::AbstractMatrix{Union{T, Nothing}},
    m_head::AbstractMatrix{Union{T, Nothing}}
) where {T<:Real}
DirectedHypergraph{T,V}(
    m_tail::AbstractMatrix{Union{T, Nothing}},
    m_head::AbstractMatrix{Union{T, Nothing}};
    v_meta::Vector{Union{Nothing,V}}=Vector{Union{Nothing,V}}(nothing, size(m,1)),
) where {T<:Real,V}
DirectedHypergraph{T,E}(
    m_tail::AbstractMatrix{Union{T, Nothing}},
    m_head::AbstractMatrix{Union{T, Nothing}};
    he_meta_tail::Vector{Union{Nothing,E}}=Vector{Union{Nothing,E}}(nothing, size(m,2)),
    he_meta_head::Vector{Union{Nothing,E}}=Vector{Union{Nothing,E}}(nothing, size(m,2))
) where {T<:Real,E}
DirectedHypergraph{T,V,E}(
    m_tail::AbstractMatrix{Union{T, Nothing}},
    m_head::AbstractMatrix{Union{T, Nothing}};
    v_meta::Vector{Union{Nothing,V}}=Vector{Union{Nothing,V}}(nothing, size(m,1)),
    he_meta_tail::Vector{Union{Nothing,E}}=Vector{Union{Nothing,E}}(nothing, size(m,2)),
    he_meta_head::Vector{Union{Nothing,E}}=Vector{Union{Nothing,E}}(nothing, size(m,2))
) where {T<:Real,V,E}
DirectedHypergraph{T,V,E,D}(
    m_tail::AbstractMatrix{Union{T, Nothing}},
    m_head::AbstractMatrix{Union{T, Nothing}};
    v_meta::Vector{Union{Nothing,V}}=Vector{Union{Nothing,V}}(nothing, size(m,1)),
    he_meta_tail::Vector{Union{Nothing,E}}=Vector{Union{Nothing,E}}(nothing, size(m,2)),
    he_meta_head::Vector{Union{Nothing,E}}=Vector{Union{Nothing,E}}(nothing, size(m,2))
) where {T<:Real,V,E,D<:AbstractDict{Int,T}}

Construct a dihypergraph using its matrix representation. In the matrix representation rows are vertices and columns are hyperedges. Optionally, values of type V can be stored at vertices and values of type E can be stored at hyperedges. By default the hypergraph uses a Dict{Int,T} for the internal data storage, however a different dictionary such as SortedDict to ensure result replicability can be used (e.g. when doing stochastic simulations on dihypergraphs).

DirectedHypergraph(g::Graphs.DiGraph)

Constructs a dihypergraph of degree 2 by making a deep copy of a Graphs.DiGraph. A SortedDict will be used for internal data storage of the dihypergraph.

DirectedHypergraph(
    hg_tail::Hypergraph{T},
    hg_head::Hypergraph{T}
) where {T<:Real}
DirectedHypergraph{T}(
    hg_tail::Hypergraph{T},
    hg_head::Hypergraph{T}
) where {T<:Real}
DirectedHypergraph{T,V}(
    hg_tail::Hypergraph{T},
    hg_head::Hypergraph{T};
    v_meta::Union{Nothing, Vector{Union{V, Nothing}}}=nothing,
) where {T<:Real,V}
 DirectedHypergraph{T,V,E}(
    hg_tail::Hypergraph{T},
    hg_head::Hypergraph{T};
    v_meta::Union{Nothing, Vector{Union{V, Nothing}}}=nothing,
    he_meta_tail::Union{Nothing, Vector{Union{E, Nothing}}}=nothing,
    he_meta_head::Union{Nothing, Vector{Union{E, Nothing}}}=nothing
) where {T<:Real,V,E}
 DirectedHypergraph{T,V,E,D}(
    hg_tail::Hypergraph{T,Nothing,Nothing,D},
    hg_head::Hypergraph{T,Nothing,Nothing,D};
    v_meta::Vector{Union{Nothing,V}}=Vector{Union{Nothing,V}}(nothing, size(hg_tail,1)),
    he_meta_tail::Vector{Union{Nothing,E}}=Vector{Union{Nothing,E}}(nothing, size(hg_tail,2)),
    he_meta_head::Vector{Union{Nothing,E}}=Vector{Union{Nothing,E}}(nothing, size(hg_tail,2))
) where {T<:Real,V,E,D<:AbstractDict{Int, T}}

Constructs a dihypergraph from two undirected hypergraphs, one with hyperedges containing "tail" vertices and one with hyperedges containing "head" verticies.

DirectedHypergraph{T,V,E,D}(
    hg_tail::Hypergraph{T,V,E,D},
    hg_head::Hypergraph{T,V,E,D}
) where {T<:Real,V,E,D<:AbstractDict{Int, T}}

Constructs a dihypergraph from two hypergraphs potentially containing metadata. Throws an error if the vertex metadata of the two hypergraphs is not element-for-element identical.

Arguments

T : type of weight values stored in the hypergraph's adjacency matrix

V : type of values stored in the vertices of the hypergraph

E : type of values stored in the edges of the hypergraph

D : dictionary for storing values the default is Dict{Int, T}

n : number of vertices

k : number of hyperedges

m : a matrix representation rows are vertices and columns are hyperedges

g : a (directed) graph representation of the hypergraph

hg_tail: an undirected hypergraph representing the tail half of the directed hypergraph

hg_head: an undirected hypergraph representing the head half of the directed hypergraph

SimpleHypergraphs.random_model — Method

random_model(
nVertices::Int,
nEdges::Int,
HType::Type{H};
no_self_loops::Bool
) where {H<:AbstractDirectedHypergraph}

Generate a random dihypergraph (in the style of Erdős–Rényi random graphs) without any structural constraints.

Algorithm

Given two integer parameters nVertices and nEdges (the number of nodes and dihyperedges, respectively), the algorithm computes - for each hyperedge e={1,...,nEdges} - two random numbers s_t ϵ [1, a] (i.e., the size of the tail) and s_h ϵ [1, b] (i.e., the size of the head). If no_self_loops is false (default), then a = b = nVertices. Otherwise, a = nVertices - 1 and b = nVertices - s_t.

Then, the algorithm selects uniformly at random s_t vertices from among [1,...,nVertices] to be added in the tail of e and s_h vertices to be added into the head of e. If no_self_loops is true (default false), then vertices will be chosen such that the same vertex v cannot appear in both the head and the tail of the same hyperedge e.

Arguments

nVertices : Number of vertices

nEdges : Number of dihyperedges

HType : Dihypergraph type (default DirectedHypergraph)

no_self_loops: If true (default false), then vertices in the tail of a dihyperedge cannot also be in the head of that same dihyperedge

Returns

A random dihypergraph of type HType with nVertices vertices and nEdges dihyperedges

SimpleHypergraphs.random_kuniform_model — Method

random_kuniform_model(
nVertices::Int,
nEdges::Int,
k::Int,
HType::Type{H};
no_self_loops::Bool
) where {H<:AbstractDirectedHypergraph}

Generates a k-uniform dihypergraph, i.e., a dihypergraph where each dihyperedge has size k = k_tail + k_head. In this implementation, k_tail and k_head are not necessarily equal.

The Algorithm

The algorithm proceeds as the random_model, forcing the size of each dihyperedge equal to k.

Arguments

nVertices : Number of vertices

nEdges : Number of dihyperedges

k : Constant total dihyperedge size

HType : Dihypergraph type (default DirectedHypergraph)

no_self_loops: If true (default false), then vertices in the tail of a dihyperedge cannot also be in the head of that same dihyperedge

Returns

A random k-uniform dihypergraph of type HType with nVertices vertices and nEdges dihyperedges

SimpleHypergraphs.random_dregular_model — Method

random_dregular_model(
    nVertices::Int,
    nEdges::Int,
    d::Int,
    HType::Type{H};
    no_self_loops::Bool=false
) where {H<:AbstractDirectedHypergraph}

Generates a d-regular directed hypergraph, where each node has degree d.

The Algorithm

The algorithm exploits the d-uniform approach described for the random_kuniform_model method to build a d-regular hypergraph h having nVertices nodes and nEdges edges. It returns the hypergraph h' dual of h.

Arguments

nVertices : Number of vertices

nEdges : Number of dihyperedges

d : Constant total dihyperedge size

HType : Dihypergraph type (default DirectedHypergraph)

no_self_loops: If true (default false), then vertices in the tail of a dihyperedge cannot also be in the head of that same dihyperedge

Returns

A random d-regular dihypergraph of type HType with nVertices vertices and nEdges dihyperedges

Manipulating vertices and hyperedges

SimpleHypergraphs.add_hyperedge! — Method

add_hyperedge!(
h::DirectedHypergraph{T, V, E, D};
    vertices_tail::D = D(),
vertices_head::D = D(),
    he_meta_tail::Union{E,Nothing}=nothing,
he_meta_head::Union{E,Nothing}=nothing
) where {T <: Real, V, E, D <: AbstractDict{Int,T}}

Adds a dihyperedge to a given dihypergraph h. Optionally, existing vertices can be added to the created dihyperedge in the tail and/or head sides. The paramater vertices_tail represents a dictionary of vertex indices and weights for the tail of the dihyperedge; vertices_head represents the vertex indices and values for the head of the dihyperedge. Additionally, metadata can be stored with the hyperedge using the he_meta_tail and he_meta_head keyword arguments.

Arguments

h : Dihypergraph

vertices_tail : Vertex index-weight dictionary with information about the vertices in the new dihyperedge's tail; default is an empty dict of type D

vertices_head : Vertex index-weight dictionary with information about the vertices in the new dihyperedge's head; default is an empty dict of type D

he_meta_tail : (Optional) metadata for the new dihyperedge's tail; default is nothing

he_meta_head : (Optional) metadata for the new dihyperedge's head; default is nothing

Returns

ix, the index of the new dihyperedge

SimpleHypergraphs.add_vertex! — Method

add_vertex!(
h::DirectedHypergraph{T, V, E, D};
    dihyperedges_tail::D = D(),
dihyperedges_head::D = D(),
v_meta::Union{V,Nothing} = nothing
) where {T <: Real, V, E, D <: AbstractDict{Int,T}}

Adds a vertex to a given dihypergraph h. Optionally, the vertex can be added to existing hyperedges. The dihyperedges_tail parameter presents a dictionary of hyperedge indices and weights stored at tail side of dihyperedges, and the dihyperedges_head parameter presents a dictionary of hyperedge identifiers and weights stored at the head side of dihyperedges. Additionally, a value can be stored with the vertex using the v_meta keyword parameter.

Arguments

h : Dihypergraph

dihyperedges_tail : Dihyperedge index-weight dictionary with information about the dihyperedges where the new vertex will be in the tail; default is an empty dict of type D

dihyperedges_head : Dihyperedge index-weight dictionary with information about the dihyperedges where the new vertex will be in the head; default is an empty dict of type D

v_meta : (Optional) metadata for the new vertex; default is nothing

Returns

ix, the index of the new vertex

SimpleHypergraphs.set_vertex_meta! — Method

set_vertex_meta!(
h::DirectedHypergraph{T, V, E, D},
new_value::Union{V,Nothing},
    v::Int
) where {T <: Real, V, E, D <: AbstractDict{Int,T}}

Sets a new metadata value new_value for the vertex v in the dihypergraph h.

Arguments

h : Dihypergraph

new_value : Metadata to replace existing metadata

v : Vertex index

Returns

h.v_meta

SimpleHypergraphs.get_vertex_meta — Method

get_vertex_meta(
h::DirectedHypergraph{T, V, E, D},
v::Int
) where {T <: Real, V, E, D <: AbstractDict{Int,T}}

Returns a metadata value associated with the vertex with index v in the dihypergraph h.

Arguments

h : Dihypergraph

v : Vertex index

Returns

Vertex metadata at index v

SimpleHypergraphs.set_hyperedge_meta! — Method

set_hyperedge_meta!(
h::DirectedHypergraph{T, V, E, D},
    new_value_tail::Union{E,Nothing},
new_value_head::Union{E,Nothing},
e::Int
) where {T <: Real, V, E, D <: AbstractDict{Int,T}}

Sets new metadata for the hyperedge with index e in the dihypergraph h.

Arguments

h : Dihypergraph

new_value_tail : Metadata value for the dihyperedge's tail (or nothing)

new_value_head : Metadata value for the dihyperedge's head (or nothing)

e : Index for the dihyperedge

Returns

(h.he_meta_tail, h.he_meta_head)

SimpleHypergraphs.set_hyperedge_meta! — Method

set_hyperedge_meta!(
h::DirectedHypergraph{T, V, E, D},
    new_value::Union{E,Nothing},
e::Int,
side::HyperedgeDirection
) where {T <: Real, V, E, D <: AbstractDict{Int,T}}

Sets a new meta value new_value for the hyperedge with index e in the direction side for a dihypergraph h.

Arguments

h : Dihypergraph

new_value : Metadata value for this dihyperedge (or nothing)

e : Dihyperedge index

side : a HyperedgeDirection (either 1 or 2)

Returns

h.he_meta_tail (if side == tail) or h.he_meta_head (if side == head)

SimpleHypergraphs.get_hyperedge_meta — Method

get_hyperedge_meta(
h::DirectedHypergraph{T, V, E, D},
e::Int
) where {T <: Real, V, E, D <: AbstractDict{Int,T}}

Returns the metadata for the dihyperedge with index e in the dihypergraph h.

Arguments

h : Dihypergraph

e : Dihyperedge index

Returns

A 2-tuple (tail_meta, head_meta), where tail_meta is the metadata associated with dihyperedge e on the tail side, and head_meta is the metadata associated with dihyperedge e on the head side.

SimpleHypergraphs.get_hyperedge_meta — Method

get_hyperedge_meta(
h::DirectedHypergraph{T, V, E, D},
e::Int,
side::HyperedgeDirection
) where {T <: Real, V, E, D <: AbstractDict{Int,T}}

Returns a metadata value for one side of the dihyperedge with index e in the dihypergraph h.

Arguments

h : Dihypergraph

e : Dihyperedge index

side : a HyperedgeDirection (either 1 or 2)

Returns

meta, where meta is the metadata associated with the tail (if side == tail) or the head (if side == head) of the dihyperedge with index e

SimpleHypergraphs.remove_vertex! — Method

remove_vertex!(h::DirectedHypergraph, v::Int)

Removes the vertex v from a given dihypergraph h. Note that running this function will cause reordering of vertices in the dihypergraph; the vertex v will replaced by the last vertex of the dihypergraph and the list of vertices will be shrunk.

Arguments

h : Dihypergraph

v : Vertex index

Returns

h

SimpleHypergraphs.remove_hyperedge! — Method

remove_hyperedge!(h::DirectedHypergraph, e::Int)

Removes the dihyperedge e from a given dihypergraph h. Note that running this function will cause reordering of hyperedges in the dihypergraph: the dihyperedge e will replaced by the last dihyperedge of the dihypergraph and the list of dihyperedges (and dihyperedge metadata) will be shrunk.

Arguments

h : Dihypergraph

e : Dihyperedge index

Returns

h

Hypergraph array getters and setters

For a directed hypergraph, an additional index can refer to either the tail (1) or head (2) of a directed hyperedge:

dh = DirectedHypergraph{Int64}(2,3);
dh[1,1,1] = 1;
dh[2,2,1] = 2;
dh

# output

2×3 DirectedHypergraph{Int64, Nothing, Nothing, Dict{Int64, Int64}}:
 (1, nothing)  (nothing, nothing)  (nothing, nothing)
 (nothing, 2)  (nothing, nothing)  (nothing, nothing)

Setting with slices is not currently possible, and users must currently use slices on the component undirected hypergraphs instead. However, accessing using slices is allowed:

dh = DirectedHypergraph{Int64}(4,2);
dh.hg_tail[1:2,1] .= 1;
dh.hg_head[3:4,1] .= 2;
dh[:,1]

# output

4-element Vector{Tuple{Union{Nothing, Int64}, Union{Nothing, Int64}}}:
 (1, nothing)
 (1, nothing)
 (nothing, 2)
 (nothing, 2)

Base.getindex — Method

Base.getindex(h::H, idx::Vararg{Int,2}) where {H <: AbstractDirectedHypergraph}

Returns a value for a given vertex-dihyperedge pair idx for a dihypergraph h. If a vertex does not belong to a dihyperedge nothing is returned.

Arguments

h : Dihypergraph

idx : Pair of integers i and j, where i is a vertex index and j is a dihyperedge index

Returns

(tail_value, head_value), where tail_value and head_value are either the weight of vertex i in the tail and head of dihyperedge j, respectively, or nothing if i is not in the tail/head.

Base.setindex! — Method

Base.setindex!(h::H, ::Nothing, idx::Vararg{Int,2}) where {H <: AbstractDirectedHypergraph}

Removes a vertex from a given hyperedge for a dihypergraph h and a given vertex-dihyperedge pair idx. Note that trying to remove a vertex from a dihyperedge when it is not present will not throw an error.

Arguments

h : Dihypergraph

nothing

idx : Pair of integers i and j, where i is a vertex index and j is a dihyperedge index

Returns

h

Base.setindex! — Method

Base.setindex!(h::H, v::Real, idx::Vararg{Int,2}) where {H <: AbstractDirectedHypergraph}

Adds a vertex to a hyperedge (represented by indices idx) and assigns value v to be stored with that assignment.

Arguments

h : Dihypergraph

v : A real-valued weight which will be applied to a vertex in both the tail and the head of a dihyperedge

idx : Pair of integers i and j, where i is a vertex index and j is a dihyperedge index

Returns

h

Base.setindex! — Method

Base.setindex!(h::H, v::Tuple{Union{Real, Nothing}, Union{Real, Nothing}}, idx::Vararg{Int,2}) where {H <: AbstractDirectedHypergraph}

Manipulates a dihyperedge (represented by indices idx), either adding a vertex to the tail and/or head sides of the hyperedge and assigning a value associated with that assignment, or else removing a vertex from the tail/head sides of the hyperedge.

Here, v is a 2-tuple where the first element is the value that will be assigned to the tail part of the hyperedge and the second element is the value that will be assigned to the head part. A value of nothing means that the vertex will be removed from that side of the hyperedge.

Arguments

h : Dihypergraph

v : A 2-tuple (tail_value, head_value). tail_value and head_value are either nothing (in which case the vertex is not in the tail or head, respectively, of the dihyperedge) or real-valued weights which will be applied to a vertex in the tail and head of a dihyperedge, respectively

idx : Pair of integers i and j, where i is a vertex index and j is a dihyperedge index

Returns

h

Base.setindex! — Method

Base.setindex!(h::H, ::Nothing, idx::Vararg{Int,3}) where {H <: AbstractDirectedHypergraph}

Removes a vertex from a given hyperedge for a dihypergraph h and a given side-vertex-dihyperedge triple idx. If the first index of idx is 1, then the vertex will be removed from the tail of the dihyperedge; if idx is 2, then the vertex will be removed from the head. Note that trying to remove a vertex from a hyperedge when it is not present will not throw an error.

Arguments

h : Dihypergraph

nothing

idx : Three-tuple of integers side, i, andj, wheresideis either1(meaning "tail") or2(meaning "head"),iis a vertex index, andj` is a dihyperedge index

Returns

h

Base.setindex! — Method

Base.setindex!(h::H, v::Real, idx::Vararg{Int,3}) where {H <: AbstractDirectedHypergraph}

Adds a vertex to a dihyperedge (represented by indices idx, where the first index must be either 1 - referring to the tail of the dihyperedge - or 2 - referring to the head) and assigns value v to be stored with that assignment.

Arguments

h : Dihypergraph

v : A real-valued weight

idx : 3-tuple of integers side, i, and j, where side is either 1 ("tail") or 2 ("head"), i is a vertex index, and j is a dihyperedge index

Returns

h

Hypergraph info

SimpleHypergraphs.nhv — Method

nhv(h::H) where {H <: AbstractDirectedHypergraph}

Return the number of vertices in the dihypergraph h.

Arguments

h : Dihypergraph

Returns

The number of vertices in h

SimpleHypergraphs.nhe — Method

nhe(h::H) where {H <: AbstractDirectedHypergraph}

Return the number of dihyperedges in the dihypergraph h.

Arguments

h : Dihypergraph

Returns

The number of dihyperedges in h. Note that if, for some reason, the he2v lists in h.hg_tail and h.hg_head have different lengths (implying that there are different numbers of dihyperedge tails than heads), an error is raised.

SimpleHypergraphs.getvertices — Method

getvertices(h::H, he_id::Int) where {H <: AbstractDirectedHypergraph}

Returns vertices (with associated weights) from a directed hypergraph h for a given hyperedge with index he_id.

Arguments

h : Dihypergraph

he_id : Dihyperedge index

Returns

2-tuple (in, out), where in contains information about the dihyperedge's tail and out contains information about the head. in and out are both AbstractDicts (following the dict (D) type in h), with the form D(v => w), where v is a vertex index and w is a real-valued weight.

SimpleHypergraphs.gethyperedges — Method

gethyperedges(h::H, v_id::Int) where {H <: AbstractDirectedHypergraph}

Returns dihyperedges (with associated weights) for a given vertex v_id in a dihypergraph h.

Arguments

h : Dihypergraph

v_id : Vertex index

Returns

2-tuple (in, out), where in contains information about the dihyperedges where vertex v_id is in the tail and out contains information about the dihyperedges where v_id is in the head. in and out are both AbstractDicts (following the dict (D) type in h), with the form D(he => w), where he is a dihyperedge index and w is a real-valued weight.

SimpleDirectedHypergraphs.get_weakly_connected_components — Method

get_weakly_connected_components(h::H) where {H <: AbstractDirectedHypergraph}

Return an array of weakly connected components in the dihypergraph h by first converting the dihypergraph into an undirected hypergraph and then obtaining the conected components of that hypergraph.

Arguments

h : Dihypergraph

Returns

A vector where each member is itself a vector consisting of the indices of one weakly connected component

SimpleDirectedHypergraphs.get_strongly_connected_components — Method

get_strongly_connected_components(h::H) where {H <: AbstractDirectedHypergraph}

Return an array of strongly connected components in the dihypergraph h, based on the "naive" algorithm of Francisco José Martín-Recuerda Moyano (PhD dissertation, 2016).

Arguments

h : Dihypergraph

Returns

A vector, where each member of the vector is itself a (sorted) vector containing the vertex indices of one strongly connected component of h

SimpleHypergraphs.get_twosection_adjacency_mx — Method

get_twosection_adjacency_mx(
h::H;
count_self_loops::Bool,
    replace_weights::Union{Nothing,Real}
) where {H<:AbstractDirectedHypergraph}

Returns an adjacency matrix for a two section view of a hypergraph h.

Note: If two vertices are connected by more than one dihyperedge in h, the associated weights of the vertices in each relevant dihyperedge will be summed.

Arguments

h : Dihypergraph

count_self_loops : If true (default false), include edges from a vertex v to v

replace_weights : If not nothing (default nothing), replace all weights with the provided value.

Returns

A m-by-m adjacency matrix, where m = nhv(h)

SimpleHypergraphs.random_walk — Method

random_walk(
    h::H,
    start::Int;
    heselect::Function,
    vselect::Function,
    reverse::bool
) where {H <: AbstractDirectedHypergraph}

Return the next vertex visited in a random walk starting from vertex start. First, a hyperedge is sampled with weights based on the output of the heselect function (by default, each hyperedge is sampled with the same probability). Next, a vertex within that dihyperedge is selected, with weights based on the output of the vselect function (by default, each vertex, including the source, is sampled with the same probability).

heselect and vselect functions take two arguments: a dihypergraph and, respectively, a vertex index or a hyperedge index. The return values of both functions should be, respectively, a list of dihyperedge or vertex indices and their weights.

Arguments

h : Dihypergraph

start : An initial vertex index

heselect : A function (see above) that assigns weights to valid dihyperedges; default assigns uniform weights to all dihyperedges

vselect : A function (see above) that assigns weights to valid vertices; default assigns uniform weights to all vertices

reverse : If true (default false), the walk will go along dihyperedges from head to tail, rather than from tail to head

Returns

A selected vertex index

SimpleHypergraphs.dual — Method

SimpleHypergraphs.dual(h::DirectedHypergraph)

Return the dual of the directed hypergraph h.

In matrix representation, the dual h' of h is the transpose. In addition to constructing a DirectedHypergraph with nhe(h) vertices and nhv(h) dihyperedges, we swap the dihyperedge and vertex metadata from h, combining the tail and head metadata for a given dihyperedge in h into a tuple (t_meta, h_meta).

NOTE: h needs to have at least one dimension greater than 0.

Arguments

h : Dihypergraph

Returns

A DirectedHypergraph that is the dual of h

Special Classes

SimpleDirectedHypergraphs.is_b_hypergraph — Function

is_b_hypergraph(h::H) where {H <: AbstractDirectedHypergraph}

A B-edge is a dihyperedge which may have multiple vertices in the tail but which has exactly one vertex in the head. A B-hypergraph is a dihypergraph where all dihyperedges are B-edges.

Arguments

h : Dihypergraph

Returns

A Bool indicating if h is a B-hypergraph (true) or not (false)

SimpleDirectedHypergraphs.is_f_hypergraph — Function

is_f_hypergraph(h::H) where {H <: AbstractDirectedHypergraph}

An F-edge is a dihyperedge which may have multiple vertices in the head but which has exactly one vertex in the tail. An F-hypergraph is a dihypergraph where all dihyperedges are F-edges.

Arguments

h : Dihypergraph

Returns

A Bool indicating if h is an F-hypergraph (true) or not (false)

SimpleDirectedHypergraphs.is_bf_hypergraph — Function

is_bf_hypergraph(h::H) where {H <: AbstractDirectedHypergraph}

A BF-hypergraph is a dihypergraph where all dihyperedges are either B-edges (see is_b_hypergraph), meaning that they have exactly one vertex in the head, or F-edges (see is_f_hypergraph), meaning that they have exactly one vertex in the tail.

Arguments

h : Dihypergraph

Returns

A Bool indicating if h is a BF-hypergraph (true) or not (false)

SimpleDirectedHypergraphs.is_loopless — Function

is_loopless(h::H; strict::Bool = true) where {H <: AbstractDirectedHypergraph}

We define a loop in two ways: "strict" and "loose".

In the "strict" definition, a dihyperedge is a loop if its tail and head contain exactly the same vertices (weights can be different). In the "loose" definition, a dihyperedge is a loop if there is any vertex that appears in both the tail and the head.

A dihypergraph is "loopless" if there are no dihyperedges that are loops.

Arguments

h : Dihypergraph

strict : Bool

Returns

loopless, a Bool indicating if h is loopless (true) or not (false)

SimpleDirectedHypergraphs.is_simple — Function

is_simple(h::H) where {H <: AbstractDirectedHypergraph}

A dihypergraph is simple if it is loopless (using the strict definition; see is_loopless) and if it is without repeated dihyperedge (meaning that there are no two hyperedges in the dihypergraph with identical tails and heads, ignoring the vertex-hyperedge weights).

Arguments

h : Dihypergraph

Returns

A Bool indicating if h is simple (true) or not (false)

SimpleDirectedHypergraphs.is_connected — Function

is_connected(h::H) where {H <: AbstractDirectedHypergraph}

A directed hypergraph is connected if it has exactly one connected component. See get_weakly_connected_components.

Arguments

h : Dihypergraph

Returns

A Bool; true if h is (weakly) connected, and false otherwise

SimpleDirectedHypergraphs.is_strongly_connected — Function

is_strongly_connected(h::H) where {H <: AbsctractDirectedHypergraphs}

A directed hypergraph is strongly connected if it has exactly one strongly connected component. See get_strongly_connected_components.

Arguments

h : Dihypergraph

Returns

A Bool; true if h is strongly connected, and false otherwise

Modifying a directed hypergraph

SimpleDirectedHypergraphs.to_undirected — Method

to_undirected(
h::DirectedHypergraph{T, V, E, D}
) where {T <: Real, V, E, D <: AbstractDict{Int, T}}

Converts a dihypergraph into an undirected hypergraph. Dihyperedge tails and heads are combined; that is, for all dihyperedges he_orig = (tail, head) in the dihypergraph h, all vertices in tail or head are added to a corresponding undirected hyperedge he_new in the undirected hypergraph h'.

Metadata is combined into tuples; i.e., if there was originally tail metadata t_meta and head metadata h_meta for a given dihyperedge he_orig, the new hyperedge he_new will have metadata (t_meta, h_meta).

Because vertex-hyperedge weights are restricted to real numbers, we cannot combine the weights, so we simply set the values to 1.0 if a given vertex is in a given hyperedge.

Arguments

h : Dihypergraph

Returns

An undirected hypergraph h'

SimpleHypergraphs.prune_hypergraph! — Method

prune_hypergraph!(h::H) where {H <: AbstractDirectedHypergraph}

Remove all vertices with degree 0 and all hyperedges of size 0.

Arguments

h : Dihypergraph

Returns

h

SimpleHypergraphs.prune_hypergraph — Method

prune_hypergraph(h::H) where {H <: AbstractDirectedHypergraph}

Remove all vertices with degree 0 and all hyperedges of size 0.

Arguments

h : Dihypergraph

Returns

A deepcopy of h with all isolated vertices and empty dihyperedges removed

Pathfinding

SimpleHypergraphs.shortest_path — Method

SimpleHypergraphs.shortest_path(
b::BipartiteView{H},
source::Int,
target::Int
) where {H<:AbstractDirectedHypergraph}

Finds a single shortest path in a graph b between vertices source and target. Note that if several paths of the same length exist, only one will be returned.

Arguments

b : BipartiteView of a dihypergraph

source : Vertex (in BipartiteView) index

target : Vertex (in BipartiteView) index

Returns

A path, represented by a sequence of edge indices

SimpleHypergraphs.shortest_path — Method

SimpleHypergraphs.shortest_path(
t::TwoSectionView{H},
source::Int,
target::Int
) where {H<:AbstractDirectedHypergraph}

Finds a single shortest path in a two-section representation of a dihypergraph t between vertices source and target. Note that if several paths of the same length exist, only one will be returned.

Arguments

t : TwoSectionView of a dihypergraph

source : Vertex index

target : Vertex index

Returns

A path, represented by a sequence of edge indices

SimpleDirectedHypergraphs.forward_reachable — Method

forward_reachable(
    h::H,
    source::Int,
) where {H <: AbstractDirectedHypergraph}

Traverses a dihypergraph h starting from vertex with index source to determine all other vertices and dihyperedges that are reachable, following dihyperedges along their forward direction (i.e., from tail to head).

Arguments

h : Dihypergraph

source : A vertex index from which to begin the search

Returns

(vs, es), where vs is a Set{Int} of reached vertex indices and es is a Set{Int} of reached dihyperedge indices

SimpleDirectedHypergraphs.backward_traceable — Method

backward_traceable(
    h::H,
    target::Int,
) where {H <: AbstractDirectedHypergraph}

Traverses a dihypergraph h starting from vertex with index target to determine all other vertices and dihyperedges that are reachable, following dihyperedges along their reverse direction (i.e., from head to tail).

Arguments

h : Dihypergraph

source : A vertex index from which to begin the search

Returns

(vs, es), where vs is a Set{Int} of reached vertex indices and es is a Set{Int} of reached dihyperedge indices

SimpleDirectedHypergraphs.is_reachable — Method

is_reachable(
    h::H,
    source::Int,
    target::Int,
    target_type::Symbol
) where {H<:AbstractDirectedHypergraph}

Use forward_reachable to determine if target (either a vertex index, if target_type === :vertex or a dihyperedge index if target_type === :dihyperedge) can be reached from vertex with index source in dihypergraph h.

Arguments

h : Dihypergraph

source : Vertex index

target : Either a vertex index or a dihyperedge index, depending on target_type

target_type : A Symbol (either :vertex or dihyperedge) indicating what type of object target represents

Returns

A Bool indicating if target is reachable from source

SimpleDirectedHypergraphs.all_hyperpaths — Function

all_hyperpaths(h::H, source::Int, target::Int) where {H <: AbstractDirectedHypergraph}

all_hyperpaths(
    h::DirectedHypergraph{T,V,E,D},
    source::Int,
    targets::Set{Int}
) where {T<:Real,V,E,D<:AbstractDict{Int,T}}

all_hyperpaths(
    h::DirectedHypergraph{T,V,E,D},
    sources::Set{Int},
    target::Int
) where {T<:Real,V,E,D<:AbstractDict{Int,T}}

all_hyperpaths(
    h::DirectedHypergraph{T,V,E,D},
    sources::Set{Int},
    targets::Set{Int}
) where {T<:Real,V,E,D<:AbstractDict{Int,T}}

Exhaustively (but efficiently) generate all hyperpaths in dihypergraph h from some source(s) to some target(s), using the algorithm described by Krieger & Kececioglu (2022), DOI: 10.1186/s13015-022-00217-9.

Note that, ostensibly, this algorithm only works for single-source, single-sink pathfinding (i.e., with a single source index and a single target index). If the user provides multiple sources and/or multiple targets, the multi-source/multi-sink problem will be reformulated as a single-source, single-sink problem by adding a metasource vertex (connected to all source vertices by a single dihyperedge) and/or metatarget vertex (connected to all target vertices by a single dihyperedge).

Arguments

h : Dihypergraph

source/sources : Either the index of a single source vertex or a Set{Int} of vertex indices

target/targets : Either the index of a single target vertex or a Set{Int} of vertex indices

Returns

paths, a Set of all possible hyperpaths from source/sources to target/targets, with each path represented as a Set{Int} of dihyperedge indices

SimpleDirectedHypergraphs.shortest_hyperpath_kk_heuristic — Function

shortest_hyperpath_kk_heuristic(
    h::H,
    source::Int,
    target::Int,
    dihyperedge_weights::AbstractVector{T}
) where {H <: AbstractDirectedHypergraph, T <: Real}

shortest_hyperpath_kk_heuristic(
    h::DirectedHypergraph{T, V, E, D},
    source::Int,
    targets::Set{Int},
    dihyperedge_weights::AbstractVector{T}
) where {T <: Real, V, E, D <: AbstractDict{Int,T}}

shortest_hyperpath_kk_heuristic(
    h::DirectedHypergraph{T, V, E, D},
    sources::Set{Int},
    target::Int,
    dihyperedge_weights::AbstractVector{T}
) where {T <: Real, V, E, D <: AbstractDict{Int,T}}

shortest_hyperpath_kk_heuristic(
    h::DirectedHypergraph{T, V, E, D},
    sources::Set{Int},
    targets::Set{Int},
    dihyperedge_weights::AbstractVector{T}
) where {T <: Real, V, E, D <: AbstractDict{Int,T}}

Implements the heuristic dihypergraph pathfinding algorithm of Krieger & Kececioglu (2022), DOI: 10.1186/s13015-022-00217-9. This algorithm is not guaranteed to find the optimal pathway from source/sources to target/targets, but in practice, it produces the optimal pathway approximately 99% of the time.

Note that, ostensibly, this algorithm only works for single-source, single-sink pathfinding (i.e., with a single source and a single target). If the user provides multiple sources and/or multiple targets, the multi-source/multi-sink problem will be reformulated as a single-source, single-sink problem by adding a metasource vertex (connected to all source vertices by a single, 0-cost dihyperedge) and/or metatarget vertex (connected to all target vertices by a single, 0-cost hyperedge).

Arguments

h : Dihypergraph

source/sources : Either the index of a single source vertex or a Set{Int} of vertex indices

target/targets : Either the index of a single target vertex or a Set{Int} of vertex indices

dihyperedge_weights : A Vector of nonnegative real-valued weights of length equal to nhe(h)

Returns

path, a Set{Int} of dihyperedge vertices constituting an estimated shortest path from source/sources to target/targets

SimpleDirectedHypergraphs.shortest_hyperpath_kk_ilp — Function

shortest_hyperpath_kk_ilp(
    h::H,
    source::Int,
    target::Int,
    dihyperedge_weights::AbstractVector{T}
) where {H<:AbstractDirectedHypergraph, T<:Real}

shortest_hyperpath_kk_ilp(
    h::DirectedHypergraph{T,V,E,D},
    source::Int,
    targets::Set{Int},
    dihyperedge_weights::AbstractVector{S}
) where {S<:Real,T<:Real,V,E,D<:AbstractDict{Int,T}}

shortest_hyperpath_kk_ilp(
    h::DirectedHypergraph{T,V,E,D},
    sources::Set{Int},
    target::Int,
    dihyperedge_weights::AbstractVector{S}
) where {S<:Real,T<:Real,V,E,D<:AbstractDict{Int,T}}

shortest_hyperpath_kk_ilp(
    h::DirectedHypergraph{T,V,E,D},
    sources::Set{Int},
    targets::Set{Int},
    dihyperedge_weights::AbstractVector{S}
) where {S<:Real,T<:Real,V,E,D<:AbstractDict{Int,T}}

Implements the exact dihypergraph pathfinding algorithm of Krieger & Kececioglu (2023), DOI: 10.1089/cmb.2023.0242. This algorithm is guaranteed to find the optimal pathway from source to target based on some nonnegative real-valued dihyperedge_weights, if and only if such a path exists.

Note that, ostensibly, this algorithm only works for single-source, single-sink pathfinding (i.e., with a single source and a single target). If the user provides multiple sources and/or multiple targets, the multi-source/multi-sink problem will be reformulated as a single-source, single-sink problem by adding a metasource vertex (connected to all source vertices by a single, 0-cost dihyperedge) and/or metatarget vertex (connected to all target vertices by a single, 0-cost dihyperedge).

Arguments

h : Dihypergraph

source/sources : Either the index of a single source vertex or a Set{Int} of vertex indices

target/targets : Either the index of a single target vertex or a Set{Int} of vertex indices

dihyperedge_weights : A Vector of nonnegative real-valued weights of length equal to nhe(h)

Returns

path, a Set{Int} of dihyperedge vertices constituting the shortest path from source/sources to target/targets

SimpleDirectedHypergraphs.SnodeDistanceKKHeuristic — Type

SnodeDistanceKKHeuristic <: AbstractDistance

Constructor: SnodeDistanceKKHeuristic(sources::Set{Int}, targets::Set{Int})

Represent a distance between a set of source nodes (sources) and a set of target nodes (targets) in a directed hypergraph, where a distance in this context is defined as the length of the shortest hyperpath that reaches all t ∈ targets.

This approach uses the heuristic shortest hyperpath algorithm of Krieger & Kececioglu 2022 (DOI: 10.1186/s13015-022-00217-9). The Krieger & Kececioglu heuristic algorithm commonly (but not always) returns the true optimal (shortest) hyperpath; this algorithm can thus be thought of as an upper bound of the true distance between the sinks and the source. To calculate the exact (minimal) distance, instead use SnodeDistanceKKExact.

Arguments

sources : A Set{Int} of vertex indices that the pathfinding algorithm starts with

targets : A Set{Int} of vertex indices that must be reached

SimpleDirectedHypergraphs.SnodeDistanceKKILP — Type

SnodeDistanceKKILP <: AbstractDistance

Constructor: SnodeDistanceKKILP(sources::Set{Int}, targets::Set{Int})

Represent a distance between a set of source nodes (sources) and a set of target nodes (targets) in a directed hypergraph, where a distance in this context is defined as the length of the shortest hyperpath that reaches all t ∈ targets.

This approach uses the exact shortest hyperpath algorithm of Krieger & Kececioglu 2023 (DOI: 10.1089/cmb.2023.0242). The Krieger & Kececioglu algorithm solves an integer linear programming formalism related to hypergraph cutting.

For a lower-cost heuristic algorithm, see SnodeDistanceKKHeuristic.

Arguments

sources : A Set{Int} of vertex indices that the pathfinding algorithm starts with

targets : A Set{Int} of vertex indices that must be reached

SimpleHypergraphs.distance — Method

distance(
    h::H,
    distance_method::SnodeDistanceKKHeuristic,
    dihyperedge_weights::AbstractVector{T};
) where {H <: AbstractDirectedHypergraph, T <: Real}

Return the shortest distance between the distance_method.sources and the distance_method.targets, assuming tha the cost (i.e., distance) between two vertices is the sum of the (nonnegative) hyperedge_weights of the hyperedges in the shortest path between the two sets of vertices. If there is no path between the sources and targets, returns typemax(T).

Here, the heuristic directed hypergraph pathfinding algorithm of Krieger & Kececioglu (2022), DOI: 10.1186/s13015-022-00217-9 is used to find the shortest path. This algorithm is not guaranteed to find the optimal pathway, but in practice, it produces the optimal pathway approximately 99% of the time.

Arguments

h : Dihypergraph

distance_method : A SnodeDistanceKKHeuristic instance

dihyperedge_weights : AbstractVector of real-valued weights with length equal to nhe(h)

Returns

If no path can be found, returns typemax(T). Otherwise, returns the sum of the weights of the dihyperedges included in the (estimated) shortest path.

SimpleHypergraphs.distance — Method

distance(
    h::H,
    distance_method::SnodeDistanceKKILP,
    dihyperedge_weights::AbstractVector{T}
) where {H <: AbstractDirectedHypergraph}

Return the shortest distance between the distance_method.sources and the distance_method.targets, assuming tha the cost (i.e., distance) between two vertices is the sum of the (nonnegative) hyperedge_weights of the hyperedges in the shortest path between the two sets of vertices. If there is no path between the sources and targets, returns typemax(T).

Here, the exact shortest hyperpath algorithm of Krieger & Kececioglu 2023 (DOI: 10.1089/cmb.2023.0242) is used. The Krieger & Kececioglu algorithm solves an integer linear programming formalism related to hypergraph cutting.

Arguments

h : Dihypergraph

distance_method : A SnodeDistanceKKILP instance

dihyperedge_weights : AbstractVector of real-valued weights with length equal to nhe(h)

Returns

If no path can be found, returns typemax(T). Otherwise, returns the sum of the weights of the dihyperedges included in the shortest path.

Graphs.diameter — Method

diameter(
    h::H,
    distance_method::SnodeDistanceKKHeuristic,
) where {H <: AbstractDirectedHypergraph, T <: Real}

Return the diameter of a hypergraph hg (maximum number of hyperedges required to go between any two vertices) based on Krieger & Kececioglu's heuristic algorithm (DOI: 10.1186/s13015-022-00217-9). If there exist some vertices not reachable from other vertices (i.e., if h is not strongly connected), then this will return Inf64.

TODO: should consider starting from more than one vertex; prerequisite problem

Arguments

h : Dihypergraph

distance_method : An SnodeDistanceKKHeuristic instance (note: initial sources and targets do not matter)

Returns

If h is not strongly connected, returns Inf64; otherwise, returns the maximum distance from any one source vertex to any one target vertex.

Graphs.diameter — Method

diameter(
    hg::H,
    distance_method::SnodeDistanceKKILP,
) where {H <: AbstractDirectedHypergraph, T <: Real}

Return the diameter of a hypergraph h (maximum number of hyperedges required to go between any two vertices) based on Krieger & Kececioglu's exact integer linear programming algorithm (DOI: 10.1089/cmb.2023.0242). If there exist some vertices not reachable from other vertices (i.e., if h is not strongly connected), then this will return Inf64.

TODO: should consider starting from more than one vertex; prerequisite problem

Arguments

h : Dihypergraph

distance_method : An SnodeDistanceKKILP instance (note: initial sources and targets do not matter)

Returns

If h is not strongly connected, returns Inf64; otherwise, returns the maximum distance from any one source vertex to any one target vertex.

Clustering

SimpleHypergraphs.quad_clustering_coefficient — Function

quad_clustering_coefficient(hg::H, v::Int) where {H <: AbstractDirectedHypergraph} 

quad_clustering_coefficient(hg::H) where {H <: AbstractDirectedHypergraph}

Implements the "quad clustering coefficient" (QCC) for directed hypergraphs, as described in: Ha, Neri, and Annibale, Chaos 34, 043102 (2024), DOI: 10.1063/5.0188246

A quad is the shortest simple cycle in a hypergraph, consisting of two vertices i and j that are both incident on the same two hyperedges α and β. The QCC is a density, describing the fraction of all possible "quads" a particular vertex i participates in. It is always true that 0 <= QCC(hg, i) <= 1.

Arguments

h : Dihypergraph

v : Vertex index

Returns

If a vertex index is given, then this function returns the QCC of that vertex. If no index is given, then instead a vector with the QCCs of all vertices (in index order) is returned.

I/O

Directed hypergraphs can be saved as and loaded from JSON- and EHGF-formatted files, where the EHGF format is a close derivative of HGF.

SimpleHypergraphs.hg_save — Function

hg_save(io::IO, h::H, format::EHGF_Format; pretty::Bool = false) where {H <: AbstractDirectedHypergraph}

Saves an undirected hypergraph h to an output stream io in ehgf format.

Arguments

io: input-output stream object

h: directed hypergraph to be saved

format: file format (here EHGF_Format, for the "extended hypergraph format")

pretty: should this be pretty-printed? (NOTE: for EHGF output, this argument does nothing, but it's still required by the SimpleHypergraphs.jl API)

hg_save(
io::IO,
h::DirectedHypergraph{T, V, E, D},
::HIF_Format;
pretty::Bool=false
) where {T, V, E, D}

Saves a directed hypergraph h to an output stream io in HIF format (see Coll et al., DOI: 10.1017/nws.2025.10018).

TODO: handling for composite metadata types

Arguments

io: input-output stream object

h: directed hypergraph to be saved

format: file format (here HIF_Format, for the "hypergraph interchange format")

pretty: should this be pretty-printed?

Returns

JSON.json output in HIF format

SimpleDirectedHypergraphs.dhg_load — Function

dhg_load(
    io::IO,
    format::EHGF_Format;
    HType::Type{H} = DirectedHypergraph,
    T::Type{U} = Bool,
    D::Type{<:AbstractDict{Int, U}} = Dict{Int, T},
) where {U <: Real, H <: AbstractDirectedHypergraph}

Loads a hypergraph from a stream io from ehgf format.

Skips a single initial comment.

Arguments

io : input-output stream object form which to load a directed hypergraph

format : file format from which to load (here, EHGF_Format, or the "extended hypergraph format")

HType : directed hypergraph datatype to store data from file (default is DirectedHypergraph)

T : real-valued datatype for incidence matrix entries/weights (default is Bool)

D : dictionary datatype for internal storage (default is Dict{Int, T})

Returns

A dihypergraph of type HType{T, Any, Any, D}

dhg_load(
io::IO,
format::HIF_Format;
HType::Type{H} = DirectedHypergraph,
T::Type{U} = Bool,
D::Type{<:AbstractDict{Int, U}} = Dict{Int, T},
V::Union{Type, Symbol} = :auto,
E::Union{Type, Symbol} = :auto,
sort_by_id::Bool = false,
add_original_id_to_meta::Union{Symbol, Nothing} = nothing,
show_warning::Bool = true
) where {H <: AbstractDirectedHypergraph, U <: Real}

Loads a directed hypergraph from an input stream io in HIF format (Coll et al., DOI: 10.1017/nws.2025.10018).

T is the type of weight values stored in the directed hypergraph matrix's matrix representation, and D is the dictionary type used to store weights for vertices in directed hyperedges. If the directed hypergraph has vertex or hyperedge metadata, their types can be specified using V and E, respectively.

Vertex and hyperedge indices are regenerated to match the 1-based indexing scheme of SimpleDirectedHypergraphs.jl. The original indices can be preserved by setting add_original_id_to_meta to a Symbol representing the key under which those indices will be stored in the metadata dictionary.

Arguments

io : input-output stream to be parsed

format : file format to be parsed (here, HIF_Format)

HType : directed hypergraph type (default is DirectedHypergraph)

T : real-valued data type for incidence weights (default is Bool)

D : dictionary data type for internal storage (default is Dict{Int, T})

V : vertex metadata type, or :auto (default is :auto)

E : hyperedge metadata type, or :auto (default is :auto)

sort_by_id : if true (default false), then the vertex and hyperedge DataFrames generated during parsing will be sorted by original ID labels

add_original_id_to_meta : ID field to be included in dataframe (to account for SimpleDirectedHypergraphs assuming sequential 1-based indexing for vertices/directed hyperedges)

show_warning : If true (default), a warning will be printed if vertex or edge IDs are changed during parsing

to_convert : symbols from the HIF attributes to be converted and included in the final dataframe

Returns

A dihypergraph of type HType with weight type T, dict type D, and vertex and dihyperedge type dependent on V, E, and the DataFrames resulting from parsing and processing

dhg_load(
    fname::AbstractString;
    format::Abstract_HG_format = EHGF_Format(),
    HType::Type{H} = DirectedHypergraph,
    T::Type{U} = Bool,
    D::Type{<:AbstractDict{Int, U}} = Dict{Int, T},
    V = Nothing,
    E = Nothing
) where {U <: Real, H <: AbstractDirectedHypergraph}

Loads a hypergraph from a file fname. The default saving format is ehgf.

Arguments

fname : filename

format : file format (default is EHGF_Format, the "extended hypergraph format")

HType : type of hypergraph to store data in

T : type of weight values stored in the hypergraph's adjacency matrix

D : dictionary for storing values the default is Dict{Int, T}

V : type of values stored in the vertices of the hypergraph

E : type of values stored in the edges of the hypergraph

Returns

A dihypergraph (of type HType in general, or of type DirectedHypergraph for EHGF-formatted inputs), with weight type T, dictionary type D, and vertex and metadata types V and E, respectively.

Note that, for parsing from HIF-formatted data, the metadata types may change

Graph utilities

Graphs.ne — Method

Graphs.ne(b::BipartiteView{H}) where {H <: AbstractDirectedHypergraph}

Arguments

b : BipartiteView of a dihypergraph

Returns

Returns the number of edges of the bipartite representation of a dihypergraph

Graphs.ne — Method

Graphs.ne(t::TwoSectionView{H}) where {H <: AbstractDirectedHypergraph}

Return the number of edges in 2-section view t of a directed hypergraph.

Arguments

t : TwoSectionView of a dihypergraph

Returns

s, the number of edges in t

Graphs.SimpleGraphs.SimpleGraph — Method

Graphs.SimpleGraph(b::BipartiteView{H}) where {H<:AbstractDirectedHypergraph}

Creates a Graphs.SimpleGraph representation of a BipartiteView b.

This creates a copy of the data. Note that the weights information is not stored in the created SimpleGraph.

Arguments

b : BipartiteView of a dihypergraph

Returns

g, a SimpleGraph

Graphs.SimpleGraphs.SimpleDiGraph — Method

Graphs.SimpleDiGraph(b::BipartiteView{H}) where {H<:AbstractDirectedHypergraph}

Creates a Graphs.SimpleDiGraph representation of a BipartiteView b.

This creates a copy of the data. Note that the weights information is not stored in the created SimpleDiGraph.

Arguments

b : BipartiteView of a dihypergraph

Returns

g, a SimpleDiGraph

Graphs.SimpleGraphs.SimpleDiGraph — Method

Graphs.SimpleDiGraph(t::TwoSectionView{H}) where {H<:AbstractDirectedHypergraph}

Creates a Graphs.SimpleDiGraph representation of a TwoSectionView t.

This creates a copy of the date. Note that the weights information is not stored in the created SimpleDiGraph.

Arguments

t : TwoSectionView of a dihypergraph

Returns

A SimpleDiGraph

Graphs.SimpleGraphs.badj — Method

Graphs.SimpleGraphs.badj(b::BipartiteView{H}) where {H<:AbstractDirectedHypergraph}

Generates a backward adjacency list for this view of a directed hypergraph.

Arguments

b : BipartiteView of a dihypergraph

Returns

For each vertex v in b, a Vector{Int} with vertex indices reachable via one edge with v as the destination

Graphs.SimpleGraphs.badj — Method

Graphs.SimpleGraphs.badj(t::TwoSectionView{H}) where {H<:AbstractDirectedHypergraph}

Generates an adjency list for this view of a hypergraph.

Graphs.SimpleGraphs.fadj — Method

Graphs.SimpleGraphs.fadj(b::BipartiteView{H}) where {H<:AbstractDirectedHypergraph}

Generates a forward adjacency list for the bipartite representation of a dihypergraph.

Arguments

b : BipartiteView of a dihypergraph

Returns

For each vertex v in b, a Vector{Int} with vertex indices reachable via one edge with v as the source

Graphs.SimpleGraphs.fadj — Method

Graphs.SimpleGraphs.fadj(t::TwoSectionView{H}) where {H<:AbstractDirectedHypergraph}

Generates a forward adjacency list for the two-section representation of a dihypergraph.

Arguments

t : TwoSectionView of a dihypergraph

Returns

For each vertex v in t, a Vector{Int} with vertex indices reachable via one edge with v as the source

Graphs.all_neighbors — Method

Graphs.all_neighbors(
    t::TwoSectionView,
    v::Integer;
    incoming::Bool = true,
outgoing::Bool = true
) where {H<:AbstractDirectedHypergraph}

Returns the indices of all vertices forward- and/or backward-adjacent to the vertex with index v in two-section view t of a dihypergraph.

Arguments

t : TwoSectionView of a dihypergraph

v : Vertex index

incoming : If true (default true), include nodes connected to v by dihyperedges where v is in the head.

outgoing : If true (default true), include nodes connected to v by dihyperedges where v is in the tail.

Returns

A Vector of vertex indices. If both incoming and outgoing are false, returns an empty Vector